Use static_cast if possible

+++ src/mip/HighsPseudocost.h
@@ -154,7 +154,7 @@ class HighsPseudocost {
 
       d = unit_gain - cost_total;
       ++nsamplestotal;
-      cost_total += d / double(nsamplestotal);
+      cost_total += d / static_cast<double>(nsamplestotal);
     } else {
       double unit_gain = -objdelta / delta;
       double d = unit_gain - pseudocostdown[col];
@@ -163,7 +163,7 @@ class HighsPseudocost {
 
       d = unit_gain - cost_total;
       ++nsamplestotal;
-      cost_total += d / double(nsamplestotal);
+      cost_total += d / static_cast<double>(nsamplestotal);
     }
   }
 
@@ -171,7 +171,7 @@ class HighsPseudocost {
                                bool upbranch) {
     double d = ninferences - inferences_total;
     ++ninferencestotal;
-    inferences_total += d / double(ninferencestotal);
+    inferences_total += d / static_cast<double>(ninferencestotal);
     if (upbranch) {
       d = ninferences - inferencesup[col];
       ninferencesup[col] += 1;
@@ -204,7 +204,7 @@ class HighsPseudocost {
     if (nsamplesup[col] == 0 || nsamplesup[col] < minreliable) {
       double weightPs = nsamplesup[col] == 0
                             ? 0
-                            : 0.9 + 0.1 * nsamplesup[col] / (double)minreliable;
+              : 0.9 + 0.1 * nsamplesup[col] / static_cast<double>(minreliable);
       cost = weightPs * pseudocostup[col];
       cost += (1.0 - weightPs) * getAvgPseudocost();
     } else
@@ -218,8 +218,7 @@ class HighsPseudocost {
 
     if (nsamplesdown[col] == 0 || nsamplesdown[col] < minreliable) {
       double weightPs = nsamplesdown[col] == 0 ? 0
-                                               : 0.9 + 0.1 * nsamplesdown[col] /
-                                                           (double)minreliable;
+                                               : 0.9 + 0.1 * nsamplesdown[col] / static_cast<double>(minreliable);
       cost = weightPs * pseudocostdown[col];
       cost += (1.0 - weightPs) * getAvgPseudocost();
     } else
@@ -256,13 +255,14 @@ class HighsPseudocost {
                             std::max(1e-6, inferences_total * inferences_total);
 
     double cutOffScoreUp =
-        ncutoffsup[col] /
-        std::max(1.0, double(ncutoffsup[col]) + double(nsamplesup[col]));
+        ncutoffsup[col] / std::max(1.0, static_cast<double>(ncutoffsup[col]) +
+                          static_cast<double>(nsamplesup[col]));
     double cutOffScoreDown =
         ncutoffsdown[col] /
-        std::max(1.0, double(ncutoffsdown[col] + nsamplesdown[col]));
-    double avgCutoffs =
-        double(ncutoffstotal) / std::max(1.0, double(ncutoffstotal + nsamplestotal));
+        std::max(1.0,
+                 static_cast<double>(ncutoffsdown[col]) + static_cast<double>(nsamplesdown[col]));
+    double avgCutoffs = static_cast<double>(ncutoffstotal) /
+        std::max(1.0, static_cast<double>(ncutoffstotal) + static_cast<double>(nsamplestotal));
 
     double cutoffScore = std::max(cutOffScoreUp, 1e-6) *
                          std::max(cutOffScoreDown, 1e-6) /
@@ -271,7 +271,8 @@ class HighsPseudocost {
     double conflictScoreUp = conflictscoreup[col] / conflict_weight;
     double conflictScoreDown = conflictscoredown[col] / conflict_weight;
     double conflictScoreAvg =
-        conflict_avg_score / (conflict_weight * double(conflictscoreup.size()));
+        conflict_avg_score /
+        (conflict_weight * static_cast<double>(conflictscoreup.size()));
     double conflictScore = std::max(conflictScoreUp, 1e-6) *
                            std::max(conflictScoreDown, 1e-6) /
                            std::max(1e-6, conflictScoreAvg * conflictScoreAvg);
@@ -297,15 +298,16 @@ class HighsPseudocost {
 
     double cutOffScoreUp =
         ncutoffsup[col] /
-        std::max(1.0, double(ncutoffsup[col] + nsamplesup[col]));
-    double avgCutoffs =
-        double(ncutoffstotal) / std::max(1.0, double(ncutoffstotal + nsamplestotal));
+        std::max(1.0, static_cast<double>(ncutoffsup[col]) + static_cast<double>(nsamplesup[col]));
+    double avgCutoffs = static_cast<double>(ncutoffstotal) /
+        std::max(1.0, static_cast<double>(ncutoffstotal) + static_cast<double>(nsamplestotal));
 
     double cutoffScore = cutOffScoreUp / std::max(1e-6, avgCutoffs);
 
     double conflictScoreUp = conflictscoreup[col] / conflict_weight;
     double conflictScoreAvg =
-        conflict_avg_score / (conflict_weight * double(conflictscoreup.size()));
+        conflict_avg_score /
+        (conflict_weight * static_cast<double>(conflictscoreup.size()));
     double conflictScore = conflictScoreUp / std::max(1e-6, conflictScoreAvg);
 
     auto mapScore = [](double score) { return 1.0 - 1.0 / (1.0 + score); };
@@ -323,15 +325,17 @@ class HighsPseudocost {
 
     double cutOffScoreDown =
         ncutoffsdown[col] /
-        std::max(1.0, double(ncutoffsdown[col] + nsamplesdown[col]));
-    double avgCutoffs =
-        double(ncutoffstotal) / std::max(1.0, double(ncutoffstotal + nsamplestotal));
+        std::max(1.0,
+                 static_cast<double>(ncutoffsdown[col]) + static_cast<double>(nsamplesdown[col]));
+    double avgCutoffs = static_cast<double>(ncutoffstotal) /
+        std::max(1.0, static_cast<double>(ncutoffstotal) + static_cast<double>(nsamplestotal));
 
     double cutoffScore = cutOffScoreDown / std::max(1e-6, avgCutoffs);
 
     double conflictScoreDown = conflictscoredown[col] / conflict_weight;
     double conflictScoreAvg =
-        conflict_avg_score / (conflict_weight * double(conflictscoredown.size()));
+        conflict_avg_score /
+        (conflict_weight * static_cast<double>(conflictscoredown.size()));
     double conflictScore = conflictScoreDown / std::max(1e-6, conflictScoreAvg);
 
     auto mapScore = [](double score) { return 1.0 - 1.0 / (1.0 + score); };
